+++ lib/health-strategies/mongodb/adapters/mongodb.js
@@ -1,15 +1,20 @@
-const Target = require('./target')
-
 /**
  * An adapter for the official node.js MongoDB driver
  * https://github.com/mongodb/node-mongodb-native
  * @class
  */
-class MongodbAdapter extends Target {
+class MongodbAdapter {
   constructor (mongodb) {
     super()
     this.client = mongodb.MongoClient
     this.connection = null
+  }
+
+  /**
+   * @return {Boolean} whether there is an active mongo connection
+   */
+  get isConnected () {
+    return this.connection !== null
   }
 
   connect (url, connectTimeoutMS) {
@@ -31,6 +36,20 @@ class MongodbAdapter extends Target {
     })
   }
 
+  /**
+   * destroys the mongodb connection
+   */
+  close () {
+    if (this.connection !== null) {
+      this.connection.close()
+      this.connection = null
+    }
+  }
+
+  /**
+   * attempts to pings mongodb
+   * @return {Promise.<Boolean>|Error}
+   */
   ping () {
     return new Promise((resolve, reject) => {
       this.connection.command({ ping: 1 }, (err) => {
